moved rapid_process.py to ecmwf_rapid_process.py
