+++ spt_ecmwf_autorapid_process/ecmwf_rapid_process.py
